KR-1-development/pipeline/requirements.txt
# Enterprise Search Pipeline Requirements
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# HTTP and API clients
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Database and vector store
astrapy==0.7.7
qdrant-client==1.7.0

# Environment and configuration
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.24.4

# Text processing and NLP
nltk==3.8.1
spacy==3.7.2
langdetect==1.0.9

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2

# Date and time utilities
python-dateutil==2.8.2
pytz==2023.3

# Async support
asyncio==3.4.3
aiofiles==23.2.1

# Logging and monitoring
loguru==0.7.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Security
cryptography==41.0.8

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Optional: ML and AI (if needed)
# torch==2.1.0
# transformers==4.35.2
# sentence-transformers==2.2.2